fix delete command for Entities. Work on adding bonuses for the ticket when admin of ship select bonuses for the ticket type

+++ src/main/java/ua/bolshak/controller/commands/DeleteCruiseCommand.java
@@ -2,6 +2,7 @@ package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.Cruise;
 import ua.bolshak.model.service.CruiseService;
+import ua.bolshak.model.service.TicketService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -12,7 +13,11 @@ public class DeleteCruiseCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         Cruise cruise = CruiseService.findById(Integer.parseInt(request.getParameter("idCruise")));
-        CruiseService.delete(cruise);
+        if (!CruiseService.checkActive(cruise)) {
+            CruiseService.delete(cruise);
+        } else {
+            request.setAttribute("ErrorMassage", "The cruise is active and can not be removed!");
+        }
         return new ToCruisesPage().execute(request, response);
     }
 }
