+++ src/main/java/ua/bolshak/controller/commands/DeleteExcursionCommand.java
@@ -2,6 +2,7 @@ package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.Excursion;
 import ua.bolshak.model.service.ExcursionService;
+import ua.bolshak.model.service.TicketService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -12,7 +13,11 @@ public class DeleteExcursionCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         Excursion excursion = ExcursionService.findById(Integer.parseInt(request.getParameter("idExcursion")));
-        ExcursionService.delete(excursion);
+        if (!TicketService.checkActiveTicker(excursion.getTickets())) {
+            ExcursionService.delete(excursion);
+        } else {
+            request.setAttribute("ErrorMassage", "The excursion has active ticket!");
+        }
         return new ToPortsPage().execute(request, response);
     }
 }
