+++ src/main/java/ua/bolshak/controller/commands/DeletePortCommand.java
@@ -1,6 +1,7 @@
 package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.Port;
+import ua.bolshak.model.service.ExcursionService;
 import ua.bolshak.model.service.PortService;
 
 import javax.servlet.ServletException;
@@ -12,7 +13,11 @@ public class DeletePortCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         Port port = PortService.findById(Integer.parseInt(request.getParameter("idPort")));
-        PortService.delete(port);
+        if (!ExcursionService.checkActiveTicker(port.getExcursions())) {
+            PortService.delete(port);
+        } else {
+            request.setAttribute("ErrorMassage", "The port has excursion");
+        }
         return new ToPortsPage().execute(request, response);
     }
 }
