+++ src/main/java/ua/bolshak/controller/commands/DeleteShipCommand.java
@@ -1,6 +1,7 @@
 package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.Ship;
+import ua.bolshak.model.service.CruiseService;
 import ua.bolshak.model.service.ShipService;
 
 import javax.servlet.ServletException;
@@ -12,7 +13,11 @@ public class DeleteShipCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         Ship ship = ShipService.findById(Integer.parseInt(request.getParameter("idShip")));
-        ShipService.delete(ship);
+        if (!CruiseService.checkActive(ship.getCruises())) {
+            ShipService.delete(ship);
+        } else {
+            request.setAttribute("ErrorMassage", "The ship has active cruise!");
+        }
         return new ToShipsPageCommand().execute(request, response);
     }
 }
