+++ src/main/java/ua/bolshak/controller/commands/DeleteShipTypeCommand.java
@@ -12,7 +12,11 @@ public class DeleteShipTypeCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         ShipType shipType = ShipTypeService.findById(Integer.parseInt(request.getParameter("ShipTypeId")));
-        ShipTypeService.delete(shipType);
+        if (!shipType.getShips().isEmpty()) {
+            ShipTypeService.delete(shipType);
+        } else {
+            request.setAttribute("ErrorMassage", "The Ship type is used!");
+        }
         return new ToShipsPageCommand().execute(request, response);
     }
 }
