+++ src/main/java/ua/bolshak/controller/commands/DeleteTicketTypeCommand.java
@@ -12,7 +12,15 @@ public class DeleteTicketTypeCommand implements ICommand {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         TicketType ticketType = TicketTypeService.findById(Integer.parseInt(request.getParameter("idTicketType")));
-        TicketTypeService.delete(ticketType);
+        if (!ticketType.getTickets().isEmpty()){
+            if (ticketType.getId() != 1) {
+                TicketTypeService.delete(ticketType);
+            } else {
+                request.setAttribute("ErrorMassage", "The standard can not be deleted!");
+            }
+        } else {
+             request.setAttribute("ErrorMassage", "The ticket type has tickets!");
+        }
         return new ToTicketsPageCommand().execute(request, response);
     }
 }
