+++ src/main/java/ua/bolshak/controller/commands/DeleteUserCommand.java
@@ -1,6 +1,7 @@
 package ua.bolshak.controller.commands;
 
 import ua.bolshak.model.entity.User;
+import ua.bolshak.model.service.TicketService;
 import ua.bolshak.model.service.UserService;
 
 import javax.servlet.ServletException;
@@ -12,8 +13,15 @@ public class DeleteUserCommand implements ICommand{
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         User user = UserService.findById(Integer.parseInt(request.getParameter("idUser")));
-
-        UserService.delete(user);
+        if (user.getRole().getId() == 2) {
+            if (!TicketService.checkActiveTicketByUser(user)) {
+                UserService.delete(user);
+            } else {
+                request.setAttribute("ErrorMassage", "The user has active tickets");
+            }
+        } else {
+            UserService.delete(user);
+        }
         return new ToUserPage().execute(request, response);
     }
 }
